+++ package.json
@@ -25,8 +25,8 @@
     "lint": "./node_modules/.bin/eslint ./packages/**/*.jsx",
     "remove-node-modules": "find . -name \"node_modules\" -exec rm -rf '{}' +",
     "lint:config": "node -e 'console.dir(require(\"./.eslintrc.js\"))'",
-    "storybook": "start-storybook -p 9001 -c packages/storybook/.storybook",
-    "deploy-storybook": "build-storybook -c packages/storybook/.storybook -o .static_storybook && storybook-to-ghpages --existing-output-dir=.static_storybook"
+    "storybook": "start-storybook -p 9001",
+    "deploy-storybook": "build-storybook -o .static_storybook && storybook-to-ghpages --existing-output-dir=.static_storybook"
   },
   "workspaces": [
     "packages/*"
@@ -58,15 +58,14 @@
     "@babel/register": "^7.9.0",
     "@commercetools/enzyme-extensions": "^5.0.0",
     "@istanbuljs/nyc-config-babel": "^3.0.0",
-    "@storybook/addon-actions": "^5.3.18",
-    "@storybook/addon-centered": "^5.3.18",
-    "@storybook/addon-docs": "^6.0.20",
-    "@storybook/addon-knobs": "^5.3.18",
-    "@storybook/addons": "^6.0.16",
-    "@storybook/cli": "^5.3.18",
-    "@storybook/react": "^5.3.18",
-    "@storybook/storybook-deployer": "^2.8.5",
-    "@storybook/theming": "^5.3.18",
+    "@storybook/addon-actions": "^6.0.26",
+    "@storybook/addon-docs": "^6.0.26",
+    "@storybook/addon-knobs": "^6.0.26",
+    "@storybook/addons": "^6.0.26",
+    "@storybook/cli": "^6.0.26",
+    "@storybook/react": "^6.0.26",
+    "@storybook/storybook-deployer": "^2.8.6",
+    "@storybook/theming": "^6.0.26",
     "@svgr/webpack": "^5.4.0",
     "@types/chai": "^4.2.12",
     "@types/classnames": "^2.2.10",
@@ -173,7 +172,9 @@
     "exclude": [
       "packages/**/*.spec.jsx",
       "packages/**/*.spec.tsx",
-      "packages/**/*.spec.ts"
+      "packages/**/*.spec.ts",
+      "packages/storybook/**",
+      "packages/utils-test/**"
     ]
   },
   "lint-staged": {
